odilia_common::cache

Struct AccessiblePrimitive

pub struct AccessiblePrimitive {
    pub sender: String,
    pub id: String,
}

A struct which represents the bare minimum of an accessible for purposes of caching. This makes some possibly eronious assumptions about what the sender is.

Fields
sender: String

Assuming that the sender is “:x.y”, this stores the (x,y) portion of this sender. Examples:

:1.1 (the first window has opened)

:2.5 (a second session exists, where at least 5 applications have been lauinched)

:1.262 (many applications have been started on this bus)

id: String

The accessible ID, which is an arbitrary string specified by the application. It is guaranteed to be unique per application. Examples:

/org/a11y/atspi/accessible/1234

/org/a11y/atspi/accessible/null

/org/a11y/atspi/accessible/root

/org/Gnome/GTK/abab22-bbbb33-2bba2

Implementations

impl AccessiblePrimitive

pub fn from_event<T: EventProperties>(event: &T) -> Self

Turns any atspi::event type into an AccessiblePrimitive, the basic type which is used for keys in the cache.

pub async fn into_accessible<'a>( self, conn: &Connection, ) -> Result<AccessibleProxy<'a>>

Convert into an atspi::proxy::accessible::AccessibleProxy. Must be async because the creation of an async proxy requires async itself.

Errors

Will return a zbus::Error in the case of an invalid destination, path, or failure to create a Proxy from those properties.