odilia_common::command

Trait TryIntoCommands

pub trait TryIntoCommands {
    type Error: Into<OdiliaError>;
    type Iter: Iterator<Item = OdiliaCommand> + Send;

    // Required method
    fn try_into_commands(self) -> Result<Self::Iter, OdiliaError>;
}

Required Associated Types

type Error: Into<OdiliaError>

type Iter: Iterator<Item = OdiliaCommand> + Send

Required Methods

fn try_into_commands(self) -> Result<Self::Iter, OdiliaError>

Fallibly returns an iterator of OdiliaCommands to run.

Errors

When implemented, the function is allowed to fail with any type that can be converted into OdiliaError, but conversion should between these types should be done from the implementers’ side, liekly using ?.