odilia_common

Module settings

Re-exports

pub use input::InputMethod;

pub use input::InputSettings;

pub use log::LogSettings;

pub use speech::SpeechSettings;

Modules

input

log

speech

Structs

ApplicationConfig

type representing a read-only view of the odilia screenreader configuration this type should only be obtained as a result of parsing odilia’s configuration files, as it containes types for each section responsible for controlling various parts of the screenreader the only way this config should change is if the configuration file changes, in which case the entire view will be replaced to reflect the fact